--- SynctreeRFC/src/main/java/com/synctree/framewok/rfc/ApplicationContextHolder.java
@@ -1,20 +1,0 @@
-package com.synctree.framewok.rfc;
-
-import org.springframework.beans.BeansException;
-import org.springframework.context.ApplicationContext;
-import org.springframework.context.ApplicationContextAware;
-import org.springframework.stereotype.Component;
-
-@Component
-public class ApplicationContextHolder implements ApplicationContextAware {
-    private static ApplicationContext context;
-
-    @Override
-    public void setApplicationContext(ApplicationContext applicationContext) throws BeansException {
-        context = applicationContext;   
-    }
-
-    public static ApplicationContext getContext() {
-        return context;
-    }
-}
