+++ SynctreeRFC/src/main/java/com/synctree/rfc/FunctionTest.java
@@ -4,7 +4,7 @@ import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.json.simple.JSONObject;
 
-import com.synctree.framewok.rfc.RFCController;
+import com.synctree.framework.rfc.RFCController;
 import com.synctree.util.Aes256Util;
 
 import java.io.UnsupportedEncodingException;
