+++ SynctreeRFC/src/main/java/com/synctree/service/InsureService.java
@@ -1,5 +1,6 @@
 package com.synctree.service;
 
+import java.sql.SQLException;
 import java.util.ArrayList;
 import java.util.HashMap;
 
@@ -9,5 +10,5 @@ public interface InsureService {
 	
 	public ArrayList<HashMap<String, String>> retrievePremium(InsureDTO dto);
 
-	public ArrayList<HashMap<String, String>> takeOutInsurance(InsureDTO dto);
+	public ArrayList<HashMap<String, String>> takeOutInsurance(InsureDTO dto) throws SQLException;
 }
