+++ SynctreeRFC/src/main/java/com/synctree/service/Temp.java
@@ -13,7 +13,7 @@ import org.springframework.stereotype.Service;
 import com.fasterxml.jackson.core.JsonParseException;
 import com.fasterxml.jackson.databind.JsonMappingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.synctree.framewok.rfc.RFCController;
+import com.synctree.framework.rfc.RFCController;
 
 @Service
 public class Temp {
